Day 7 Part 1: Score a game of cards
Wnet a little enum heavy on this so we could do the whole ordering of
hands using PartialOrd traits.

Adds in assert2 crate for the check! macro, so we can assert things and
have them pretty print nicely.

Adds in test_case crate so we can do some more table style testing for
simple stuff (like is this recognised as 4 of a kind)