Day 07 Part 2: Joker's Wild
In order to allow Jokers to be Wild we need to stop doing some of the
more succint from_str() stuff so we can construct the hand using the
modified rules.  We should also consider renaming away from from_str so
it's more apparent we're using a constructor.

Luckily we don't need to get too fancy with actually making jokers wild,
all we need to do is to add the number of Jacks to the highest card, and
that's always the better hand (see back of napkin for scribblings when I
though that may not be the case).  The other thing is to add a Joker
value, and when jokers_wild we swap to that (lower) value.