Most of Day 10 Part 2
We go round the loop, flooding on the right side to find what we are the
boundary for.

This works half the time, but when the path is anti-clockwise it ends up
flooding off to infinity, or rather it stack overflows because the
recursion is trying to do that.

Need to come up with some way to ensure we have a clockwise path.  More
rights than lefts?