Day 10 Part 2: complete and working
We needed to ensure we always walked our path clockwise, so we can rely
on 'touching' the right wall to start flood_filling to find the enclosed
cells.

Also needed to figure out the cases for when you round a corner, as to
what cells to consider as adjacent, as they weren't all by the nice and
simple vertical and horizontal cases.

This one's been the most coded one so far.  Need to take a look at
refactoring it down with fresh eyes.