Day 11 Part 1 complete
Add in slow_rows and slow_cols, these are vectors that match up to the
expansion zones, and so are summed into the computed distances.

A little extra makework to order the rows and columns so we know which
is the smallest - std::cmp::minmax would be useful here, but's it's only
in nightly.

https://doc.rust-lang.org/nightly/std/cmp/fn.minmax.html