Day 13 Part 2: Smudged matching of images
Add a new property to an image to indicate that it's smudged.  The
matching of a smudged mirror line to a non-smudged is the difference of
just one mismatched comparison in all the lines we look at.

This does need us to unfold the is_horizontal_mirror test, so now it's
equally ugly when compared to the is_vertical_mirror test, but apart
from threading the variable in, it's not much worst than the part 1
code.