Day 14 Part 4: Do something slow 1_000_000_000 times
Another AOC classic.  We need to know what the state of the platform
will be after 1_000_000_000 spins.  Given it takes about 2 seconds to
fully spin the platform around 100 times (where we find our repeat), we
need to find a shortcut, and that's repeating cycles.

Run the simulation and cache the state and the loop on which
we saw it, until we get a repeated state.  We then compute how long this
cycle is (step), the number of steps we need to complete to reach our
target (steps_left), and then advance the loop counter to that value and
simulate the few remaining cycles.

Also adds pretty-printing for the Rocks state, so we could visually
check that roll_east/roll_west were going in the correct direction.