Day 15 Part 2: Maintain the HASHMAP and report on it
Tweaked the hoho_hash to take &str and return usize; shouldn't hurt
too much and is a little more ergonomic to write the rest with as
there's a lot of using the hash as an index.

For the rest, simply follow the instructions and build up a structure.
Making an array of Vec was oddly most of the work:

    let mut boxes: [Vec<Lens>; 256] = [(); 256].map(|()| Vec::new());

That's one of those dense lines that python folks who think array
comprehensions are the best thing since sliced bread love.

Would have preferred to have written it more like so:

    let mut boxes: [Vec<Lens>; 256] = Default::default();

but that's only supported for N <= 32, as it's actually expanded by hand
in the library code.