Day 16 Part 2: Bring out the brute force
Make energised take a single parameter of the starting beam, and then
generate all the values and select the max.

This could be improved with a shared cache, as given a beam vector we
could memoise how many cells it energises, but it's quick enough just
running the previous solution loop 2*(width+height) times.

Part 1: 7434 (2.2ms)
Part 2: 8183 (481.1ms)