Day 17: Custom walk of a maze
Use djikstra from the pathfinding create.

Brings in the strum crate so we can be lazy and iterate all the
directions.

Our neighbours function/state (Cart) tracks what direction we're moving in, and
how long we have been, so we can simply drop runs too long, or U-turns,
in addition to the steps that would be out of bounds.