Day 18 Part 2: Derp a derp
I ssumed we'd be doing somthing fun with calculating colors,  it turns
out that the color data is actually movement, so we end up with quite a
lot of moving around.

This means we can't reasonably use the approach of flood filling an area
of a path, and instead will need to so something with the coordinates of
the path to compute rectangles, or something.

Hre is the test for the hex decoding, and a miswired (and failing) test
for part 2.  Time to grab some paper and see what we can do to calculate
the area of rectangles from a stream of points.