Day 19 Part 2: Figure out accepted ranges
We introduce a ProductMinMax which looks a little like a Product but
its values are MinMax's that default to the range of each rating
(1..4000)

Then we just walk down the tree from "in" but instead of actually
matching the rules, we update the possible ranges for (a) the branch (b)
the fallthrough, narrowing based on the operands we see.

When we hit an accept state, we can ask the corresponding ProductMinMax for its
possible_combinations, which is just the product of all the attributes ranges.