Day 20 Part 2: More cycle detection
By looking at rx in the graph it has one parent, this parent
(`rx_source` in our code) is a conjunction, which has 4 modules
connected to it.

Since we know `rx_source` won't go low until all the connected modules
go high, and we also know we're doing AOC, we press the button until we
see the target of `rx_source` and the value Pulse::High go past, and
record the cycle.  Once we have all the cycles, it'll be the LCM of them
all to find out when the machine will reboot (our answer).

We don't have tests for this, as there is no sample input with the rx
tree, but we do make some assertions about the structure of the network
as we go.