Day 21 Part 2: For reals
Had to go to reddit for the cluebat here.

If you look at the input rather than the example, you see that
there's a clear path from the start (always the center) to the edge or
the corner of the garden, which means it'll grow up to be a nice clean
polynomial, with a period of the width being significant.

This means we can compute the first 3 steps, derive the polynomial
interpolation, and then from those values compute the result.

This could be a little more efficient in finding the first three values,
if we made say Garden.steps take and return an array of values, but it's
not so bad.