Day 22 Part 1: Simple simulation
Load our input into a Pile of Brick.  Pile holds the adjaceny lists of
what brick sits on what brick, and what brick is supported by that.
It's populated by Pile.drop() which is the simple simulation of dropping
all the bricks to 0, lowest first.

Once we've dropped that, our first answer is find the bricks that have
multiple supporting blocks.