Day 23 Part 1:  Try and do a max flow calculation
This looks like it might be solved by a maximum flowrate algorithm, as
the prettiest path is the one that picks up the most nodes.

We're using the pathfinding crate's Matrix here, as it saves a tiny
amount of code in doing neighbour calcuations, as it'll never suggest a
neighbour that's out of bounds.

Our neighbour code though does need to know that you can't go directly
up a hill, so we add on the filter for that and the map back into our
Pos type and return the iterator on.

Not currently able to make the edmonds_karp calculation give me
anything useful, so it might be the wrong tool for the job, may be
better of just doing a BFS with a distance HashMap.