Day 23 Part 2: Spadework
So it turns out our basic brute force doesn't work with the full test
input.

Here we've split up generating a graph from the maze from walking the
graph to find the resulting paths.

We've also started on a function Graph.collapse() which should take any
runs of 2 nodes and fold them down.  It's not working, so it'll need
some tests written.