Day 24 Part 1: Line intersections, and it happens in the future.
Oh maths lesson.

The key things we do here are translate the lines from point,vector form
into y=ax + b  so we can solve for the intersection point.  Then we need
to look at the vectors and just determine which side of the origin they
point (in_future)

I'm not looking forward to part 2 where I guess we need to solve the
system of equations for some new constraint.