Day 24 Part 2: Bring out the big guns
So it's a system of equations we need to solve, so bring in z3 to do the
heavy lifting.

I'm sure in true AOC style there's some neat analysis we could do to the
data to generate an solution, and looking at the reddit there's a
big split between those who did that, those who just used z3.

Build: on linux need to ensure `libz3-dev` and `libclang-dev` are installed.