Day 25 Scaffold and borrow checker fight
Attempt at a brute force partition.

Assume we can greedily take the most connected node out from the graph
and add it to the graph on the left.  Repeat this until the number of
remaining edges between the two graphs is 3, then we can count up.

The problem though is we're in a fight over the mutable/immutable
references to `left`.  The closure and loop condition are immutable
references, but we use a mutable one at the tail of the loop to remove
the most connected node.